backend/requirements.txt
fastapi
uvicorn
youtube-transcript-api
groq
python-dotenv
pydantic
